Add Adder, Divider

+++ app/src/main/java/codylab/cwliu/tryrobo/Adder.java
@@ -2,7 +2,7 @@ package codylab.cwliu.tryrobo;
 
 public class Adder {
 
-    public static int add(int x, int y){
-        return x+y;
+    public static int add(double x, double y) {
+        return (int) x + (int) y;
     }
 }
